requirements.txt
asgiref==3.6.0
django-anymail==9.0
Django==4.1.7
djangorestframework==3.14.0
drf-spectacular==0.25.1
pytz==2022.7.1
sqlparse==0.4.3
